move test helpers to dedicated module
