+++ tests/physical_models/crop/test_assimilation.py
@@ -4,11 +4,11 @@ import torch
 from pcse.models import Wofost72_PP
 from diffwofost.physical_models.config import Configuration
 from diffwofost.physical_models.crop.assimilation import WOFOST72_Assimilation
-from diffwofost.physical_models.utils import EngineTestHelper
+from diffwofost.physical_models.test import EngineTestHelper
+from diffwofost.physical_models.test import calculate_numerical_grad
+from diffwofost.physical_models.test import get_test_data
+from diffwofost.physical_models.test import prepare_engine_input
 from diffwofost.physical_models.utils import _afgen_y_mask
-from diffwofost.physical_models.utils import calculate_numerical_grad
-from diffwofost.physical_models.utils import get_test_data
-from diffwofost.physical_models.utils import prepare_engine_input
 from .. import phy_data_folder
 
 assimilation_config = Configuration(
