+++ tests/physical_models/soil/test_waterbalance.py
@@ -8,10 +8,10 @@ from diffwofost.physical_models.config import Configuration
 from diffwofost.physical_models.crop.wofost72 import Wofost72
 from diffwofost.physical_models.soil.classic_waterbalance import WaterbalanceFD
 from diffwofost.physical_models.soil.classic_waterbalance import WaterbalancePP
-from diffwofost.physical_models.utils import EngineTestHelper
-from diffwofost.physical_models.utils import calculate_numerical_grad
-from diffwofost.physical_models.utils import get_test_data
-from diffwofost.physical_models.utils import prepare_engine_input
+from diffwofost.physical_models.test import EngineTestHelper
+from diffwofost.physical_models.test import calculate_numerical_grad
+from diffwofost.physical_models.test import get_test_data
+from diffwofost.physical_models.test import prepare_engine_input
 from .. import phy_data_folder
 
 waterbalance_config = Configuration(
